Refactor package structure for metadata extraction classes in spring-content-commons module.
